chore: adjust config overrides, delete redundant helpers
