+++ packages/core/src/index.ts
@@ -32,7 +32,7 @@ export {
 export type { ToolExecutionContext, OperationType, ToolCategory, CallToolResult } from "@mongodb-js/mcp-types";
 export { ReactiveResource } from "./reactiveResource.js";
 export type { AnyResourceClass, ResourceClass } from "@mongodb-js/mcp-types";
-export { getRandomUUID } from "./randomUUID.js";
+export { getRandomUUID } from "./getRandomUUID.js";
 export { TRANSPORT_PAYLOAD_LIMITS } from "./transportConstants.js";
 export { CommonArgs, NO_UNICODE_ERROR } from "./args.js";
 export { LogId } from "./logId.js";
