+++ packages/integration-tests/package.json
@@ -36,6 +36,7 @@
     "@mongodb-js/mcp-tools-atlas": "workspace:*",
     "@mongodb-js/mcp-setup": "workspace:*",
     "@mongodb-js/mcp-tools-atlas-local": "workspace:*",
+    "@mongodb-js/mcp-tools-assistant": "workspace:*",
     "@mongodb-js/mcp-http-runners": "workspace:*",
     "@mongodb-js/devtools-proxy-support": "^0.7.8",
     "@mongodb-js/mcp-cli": "workspace:*",
