+++ packages/integration-tests/src/transports/configOverrides.test.ts
@@ -15,15 +15,9 @@ import type {
 import { CompositeLogger, Keychain } from "@mongodb-js/mcp-core";
 import type { StreamableHTTPServerTransport } from "@modelcontextprotocol/sdk/server/streamableHttp.js";
 import { Server } from "mongodb-mcp-server";
-import { AllTools } from "mongodb-mcp-server/tools";
-import {
-    Session,
-    Elicitation,
-    connectionErrorHandler,
-    MCPConnectionManager,
-    ExportsManager,
-    applyConfigOverrides,
-} from "mongodb-mcp-server";
+import { AllTools } from "mongodb-mcp-server";
+import { applyConfigOverrides } from "@mongodb-js/mcp-cli";
+import { Session, Elicitation, connectionErrorHandler, MCPConnectionManager, ExportsManager } from "mongodb-mcp-server";
 import { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
 import { ApiClient } from "@mongodb-js/mcp-atlas-api-client";
 import { createAtlasLocalClient } from "@mongodb-js/mcp-tools-atlas-local";
