+++ packages/integration-tests/src/transports/streamableHttp.test.ts
@@ -29,7 +29,7 @@ import {
     ExportsManager,
     packageInfo,
 } from "mongodb-mcp-server";
-import { AllTools } from "mongodb-mcp-server/tools";
+import { AllTools } from "mongodb-mcp-server";
 import type { StreamableHTTPServerTransport } from "@modelcontextprotocol/sdk/server/streamableHttp.js";
 import type { CallToolResult } from "@mongodb-js/mcp-types";
 import type { AtlasTelemetry, TelemetryToolMetadata } from "@mongodb-js/mcp-atlas-telemetry";
