+++ packages/integration-tests/src/ui/mcpUIFeature.test.ts
@@ -9,7 +9,7 @@ import {
 import { CompositeLogger } from "@mongodb-js/mcp-core";
 import { ExportsManager } from "@mongodb-js/mcp-tools-mongodb";
 import { Session } from "mongodb-mcp-server";
-import { AllTools } from "mongodb-mcp-server/tools";
+import { AllTools } from "mongodb-mcp-server";
 import { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
 import { Server } from "mongodb-mcp-server";
 import { MCPConnectionManager } from "@mongodb-js/mcp-tools-mongodb";
