+++ packages/mongodb-mcp-server/src/tools/index.ts
@@ -1,24 +1,12 @@
-import { AtlasTools } from "@mongodb-js/mcp-tools-atlas";
-import { AtlasLocalTools } from "@mongodb-js/mcp-tools-atlas-local";
-import { MongoDBTools } from "@mongodb-js/mcp-tools-mongodb";
-import { AssistantTools } from "@mongodb-js/mcp-tools-assistant";
-import type { ToolClass } from "@mongodb-js/mcp-core";
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export const AllTools: ToolClass<any>[] = [
-    ...MongoDBTools,
-    ...AtlasTools,
-    ...AtlasLocalTools,
-    ...AssistantTools,
-] as const;
-
+/**
+ * Temporary file for comparing the API Extractor report only (`api-extractor/api-extractor-tools.json`).
+ * Not published as a package export — import from `@mongodb-js/mcp-tools-*` or `mongodb-mcp-server`.
+ */
 export { MongoDBToolBase, type IMongoDBConfig, type IMongoDBSession } from "@mongodb-js/mcp-tools-mongodb";
 
-// Export all the individual tools for handpicking
 export * from "@mongodb-js/mcp-tools-atlas-local";
 export * from "@mongodb-js/mcp-tools-mongodb";
 export * from "@mongodb-js/mcp-tools-assistant";
-// Named atlas exports (avoid star-export duplication with mongodb tools package).
 export { AtlasTools, AtlasToolBase, type IAtlasConfig, type IAtlasSession } from "@mongodb-js/mcp-tools-atlas";
 
 export {
@@ -30,4 +18,4 @@ export {
     type ToolArgs,
     type ToolExecutionContext,
     type ToolResult,
-} from "./tool.js";
+} from "@mongodb-js/mcp-core";
