--- packages/mongodb-mcp-server/src/tools/tool.ts
@@ -1,13 +1,0 @@
-export {
-    ToolBase,
-    type ToolClass,
-    type AnyToolClass,
-    type ToolConstructorParams,
-    type AnyToolBase,
-    type ToolArgs,
-    type ToolExecutionContext,
-    type ToolResult,
-    type OperationType,
-    type ToolCategory,
-    formatUntrustedData,
-} from "@mongodb-js/mcp-core";
